+++ fe/src/components/onboarding/ProfileForm.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { z } from 'zod';
 import { useForm, useWatch } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 
@@ -9,25 +8,15 @@ import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 
-const profileSchema = z.object({
-  nickname: z.string().trim().min(1, '닉네임을 입력이 필요합니다.'),
-  birth_date: z
-    .string()
-    .trim()
-    .min(1, '생년월일 입력이 필요합니다.')
-    .regex(/^\d{4}-\d{2}-\d{2}$/, 'YYYY-MM-DD 형식으로 입력해주세요.'),
-  gender: z.enum(['male', 'female']),
-});
-
-export type ProfileFormValues = z.infer<typeof profileSchema>;
+import { profileSchema, type ProfileFormValues } from '@/schemas/profile';
 
 type ProfileFormProps = {
-  isSubmitting: boolean;
+  defaultValues?: Partial<ProfileFormValues>;
   onSubmit: (data: ProfileFormValues) => void;
 };
 
 export default function ProfileForm({
-  isSubmitting,
+  defaultValues,
   onSubmit,
 }: ProfileFormProps) {
   const {
@@ -35,15 +24,11 @@ export default function ProfileForm({
     handleSubmit,
     setValue,
     control,
-    formState: { errors },
+    formState: { errors, isSubmitting, isDirty },
   } = useForm<ProfileFormValues>({
     resolver: zodResolver(profileSchema),
     mode: 'onSubmit',
-    defaultValues: {
-      nickname: '',
-      birth_date: '',
-      gender: 'male',
-    },
+    defaultValues,
   });
 
   const gender = useWatch({ control, name: 'gender' });
@@ -102,7 +87,11 @@ export default function ProfileForm({
         ) : null}
       </div>
 
-      <Button type="submit" className="w-full" disabled={isSubmitting}>
+      <Button
+        type="submit"
+        className="w-full"
+        disabled={isSubmitting || !isDirty}
+      >
         저장
       </Button>
     </form>
